SvgUtils: add namespace-aware filtering functions
Add helper functions that can operate on xml svg data containing the
usual svg namespace. That is, the entry name of svg nodes is always
prefixed by a huge namespace + the entry itself, for instance:

`http://www.w3.org/2000/svg:path`

Would be the tag of the "path" entry.

This patch adds functions for retrieving paths and groups that are
direct children of a given xml node.