animator: add animation for circle outlines
For circles, the "fill" property becomes annoying. While the outline can be
animated in the usual way, the fill property being set to 1 makes the animation
useless. We could animate the fill property too, but it wouldn't be  natural.

It would be ideal if could animate the "fill %", i.e., how full the circle is.

As it stands, I think the recommendation is to not use circles.