animator: [nfc] refactor keyboard events
Instead of having a state named "listen_to_kb", we call it "ongoing_animation"
to better reflect what is going on on the program. This allows a more clear
logic about what should be done in program. For instance, an input for "next"
should reason about whether there is an ongoing animation:

- If an animation is in progress, finish it in the next frame (TBD).
- If no animation is in progress, start the next animation.

The animator should always be listening to the inputs.